requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
cmsplugin-filer==1.1.3
csc-pysparse==1.1.1.4
csc-utils==0.6.7
decorator==4.0.11
Divisi2==2.2.5
dj-database-url==0.4.2
Django==1.8.18
django-appconf==1.0.2
django-autoslug-field==0.2.3
django-braces==1.11.0
django-classy-tags==0.8.0
django-cms==3.4.2
django-filer==1.2.7
django-filer-pdf==0.0.1
django-formtools==2.0
django-mptt==0.8.7
django-polymorphic==1.0.2
django-recommends==0.4.0
django-sekizai==0.10.0
django-treebeard==4.1.0
djangocms-admin-style==1.2.6.2
djangocms-attributes-field==0.2.0
djangocms-column==1.7.0
djangocms-googlemap==1.0.2
djangocms-installer==0.9.6
djangocms-link==2.1.1
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
easy-thumbnails==2.3
html5lib==0.9999999
kombu==4.0.2
lightfm==1.12
networkx==1.11
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
pyPdf==1.13
python-dateutil==2.6.0
python-recsys==0.2
pytz==2017.2
requests==2.13.0
scipy==0.19.0
six==1.10.0
South==1.0.2
tzlocal==1.3
Unidecode==0.4.20
vine==1.1.3
Wand==0.4.4
